dev-requirements.txt
-r requirements.txt
pytest==5.2.2
pytest-django==3.6.0
pytest-mock==3.2.0
